+++ src/dlt_filesystem/source/fsspec/webhdfs.py
@@ -2,6 +2,7 @@ from typing import Type
 
 from fsspec import AbstractFileSystem
 
+from dlt_filesystem.error import MissingConnectorOption
 from dlt_filesystem.source.base import FilesystemSource
 from dlt_filesystem.source.core import infer_resource
 from dlt_filesystem.source.model import FilesystemLocator
@@ -49,6 +50,10 @@ class WebHDFSSource(FilesystemSource):
         cast_to_int(fs_kwargs, ["port"])
         cast_to_bool(fs_kwargs, ["kerberos", "use_https", "session_verify"])
         cast_to_dict(fs_kwargs, ["kerb_kwargs", "data_proxy"])
+
+        # Sanity checks.
+        if "host" not in fs_kwargs or not fs_kwargs["host"]:
+            raise MissingConnectorOption("host", "WebHDFS")
 
         # Create filesystem and dlt resource wrapper.
         fs = self.fs_class(**fs_kwargs)