+++ docs/changelog.md
@@ -3,8 +3,9 @@
 ## in progress
 
 - Filesystem: Migrated local filesystem access to Apache Arrow.
-- Filesystem: Added support for reading from
-  Dropbox, FTP, HDFS, OCI, OneDrive, OSS, R2, SharePoint, SMB, WebDAV.
+- Filesystem: Added support for reading from Databricks, Dropbox,
+  FTP, Google Drive, HDFS, OCI, OneDrive, OSS, R2, SharePoint,
+  SMB, WebDAV, and WebHDFS.
 
 ## 2026/07/16 v0.7.0